Refactor code with Java project structure for improved readability
